Type casting corrections
VS throws errors in several locations about ambigious type casting.
Fixed by removing the ambiguities and casting things to double.

+++ Source/Processors/Visualization/SpikeObject.cpp
@@ -202,7 +202,7 @@ void generateSimulatedSpike(SpikeObject *s, uint64_t timestamp, int noise)
 				n = rand() % noise - noise/2;
 			}
 
-            s->data[idx] = (trace[waveType][j] + n)  * pow(scale[j],scaleExponent) + shift;
+            s->data[idx] = (trace[waveType][j] + n)  * pow(double(scale[j]),scaleExponent) + shift;
             idx = idx+1;
         }
     }
